refactor(prolong): rename scripts
