+++ servers/republik/script/prolong/mcRegular.js
@@ -1,12 +1,11 @@
 #!/usr/bin/env node
 /**
  * This script adds to mailchimp members:
- * - a CUSTOM_PLEDGE if they need to prolong before 2019-01-16
- * - the BONUS tag for who is eligable
+ * - a CP_ATOKEN2 if they need to prolong before 2019-01-16
  *
  *
  * Usage: (run from servers/republik)
- * cat local/cancellations.csv | node script/prolong/updateMailchimp.js
+ * node script/prolong/mcRegular.js [--dry]
  */
 require('@orbiting/backend-modules-env').config()
 const PgDb = require('@orbiting/backend-modules-base/lib/pgdb')
@@ -15,9 +14,9 @@ const crypto = require('crypto')
 const sleep = require('await-sleep')
 const fetch = require('isomorphic-unfetch')
 const moment = require('moment')
+const uniqBy = require('lodash/uniqBy')
 const { transformUser, AccessToken } = require('@orbiting/backend-modules-auth')
 const {
-  isBonusEligable: getIsBonusEligable,
   prolongBeforeDate: getProlongBeforeDate
 } = require('../../modules/crowdfundings/graphql/resolvers/User')
 
@@ -34,7 +33,7 @@ if (!MAILCHIMP_MAIN_LIST_ID) {
 
 const STATS_INTERVAL_SECS = 2
 const PROLONG_BEFORE_DATE = moment('2019-01-16')
-const BONUS_TAG_NAME = 'BONUS'
+const TOKEN_FIELD = 'CP_ATOKEN2'
 
 const me = {
   roles: ['admin']
@@ -67,27 +66,15 @@ PgDb.connect().then(async pgdb => {
     console.log("dry run: this won't change anything")
   }
 
-  // load users with membership or pledge
+  // load users with a membership
   const users = await pgdb.query(`
     SELECT
-      u.*
+      DISTINCT(u.*)
     FROM
       users u
     JOIN
       memberships m
       ON m."userId" = u.id
-    WHERE
-      u.id != :PARKING_USER_ID
-
-    UNION
-
-    SELECT
-      u.*
-    FROM
-      users u
-    JOIN
-      pledges p
-      ON p."userId" = u.id
     WHERE
       u.id != :PARKING_USER_ID
   `, {
@@ -99,9 +86,7 @@ PgDb.connect().then(async pgdb => {
   console.log(`investigating ${users.length} users`)
 
   const stats = {
-    numBonusEligable: 0,
     numNeedProlong: 0,
-    numBonusEligableProgress: 0,
     numNeedProlongProgress: 0,
     numOperations: 0
   }
@@ -112,7 +97,11 @@ PgDb.connect().then(async pgdb => {
   const inNeedForProlongUsers = await Promise.filter(
     users,
     async (user) => {
-      const prolongBeforeDate = await getProlongBeforeDate(user, {}, { pgdb, user: me })
+      const prolongBeforeDate = await getProlongBeforeDate(
+        user,
+        { ignoreClaimedMemberships: true },
+        { pgdb, user: me }
+      )
       stats.numNeedProlongProgress += 1
       if (prolongBeforeDate && moment(prolongBeforeDate).isBefore(PROLONG_BEFORE_DATE)) {
         stats.numNeedProlong += 1
@@ -124,23 +113,83 @@ PgDb.connect().then(async pgdb => {
   )
   delete stats.numNeedProlongProgress
 
-  const bonusEligableUsers = await Promise.filter(
-    inNeedForProlongUsers,
-    async (user) => {
-      const eligable = await getIsBonusEligable(user, {}, { pgdb, user: me })
-      stats.numBonusEligableProgress += 1
-      if (eligable) {
-        stats.numBonusEligable += 1
-      }
-      return eligable
-    },
-    {concurrency: 10}
+  // users who gifted a membership, which is active,
+  // renewing and needs prolong before PROLONG_BEFORE_DATE
+  const canProlongUsers = await pgdb.query(`
+    WITH dormant_memberships AS (
+      SELECT
+        m.*
+      FROM
+        memberships m
+      JOIN
+        pledges p
+        ON m."pledgeId" = p.id
+      JOIN
+        packages pkg
+        ON p."packageId" = pkg.id
+      WHERE
+        m.id NOT IN (SELECT "membershipId" FROM "membershipPeriods") AND --no period
+        pkg.name != 'ABO_GIVE' AND
+        m."userId" != :PARKING_USER_ID AND
+        m."membershipTypeId" IN (
+          SELECT id FROM "membershipTypes" WHERE name = ANY('{ABO, BENEFACTOR_ABO}')
+        )
+    ), givers AS (
+      SELECT
+        u.id AS "userId",
+        m.id AS "membershipId",
+        max(mp."endDate") AS "maxEndDate"
+      FROM
+        users u
+      JOIN
+        pledges p
+        ON p."userId" = u.id
+      JOIN
+        memberships m
+        ON
+          m."pledgeId" = p.id AND
+          m."userId" != u.id AND
+          m."active" = true AND
+          m."renew" = true
+      JOIN
+        "membershipPeriods" mp
+        ON
+          m.id = mp."membershipId"
+      WHERE
+        u.id != :PARKING_USER_ID AND
+        u.id NOT IN (SELECT DISTINCT("userId") FROM dormant_memberships)
+      GROUP BY
+        1, 2
+      ORDER BY
+       1, 3
+    )
+      SELECT
+        u.*
+      FROM
+        givers
+      JOIN
+        users u
+        ON givers."userId" = u.id
+      WHERE
+        "maxEndDate" < :PROLONG_BEFORE_DATE
+  `, {
+    PARKING_USER_ID,
+    PROLONG_BEFORE_DATE
+  })
+  stats.numCanProlongUsers = canProlongUsers.length
+
+  const prolongUsers = uniqBy(
+    [
+      ...inNeedForProlongUsers,
+      ...canProlongUsers
+    ],
+    u => u.id
   )
-  delete stats.numBonusEligableProgress
+  stats.numProlongUsers = prolongUsers.length
 
   console.log('building operations...')
-  let operations = await Promise.map(
-    inNeedForProlongUsers,
+  const operations = await Promise.map(
+    prolongUsers,
     async (user) => {
       const email = user.email.toLowerCase()
       const accessToken = await AccessToken.generateForUser(user, 'CUSTOM_PLEDGE')
@@ -153,31 +202,12 @@ PgDb.connect().then(async pgdb => {
           email_address: email,
           status_if_new: 'subscribed',
           'merge_fields': {
-            'CP_ATOKEN': accessToken
+            [TOKEN_FIELD]: accessToken
           }
-          // of course tags can not be provided here 🤬
         })
       }
     },
     {concurrency: 10}
-  )
-  operations = operations.concat(
-    bonusEligableUsers.map(user => {
-      const email = user.email.toLowerCase()
-      stats.numOperations += 1
-      return {
-        method: 'POST',
-        path: `/lists/${MAILCHIMP_MAIN_LIST_ID}/members/${hash(email)}/tags`,
-        body: JSON.stringify({
-          tags: [
-            {
-              name: BONUS_TAG_NAME,
-              status: 'active'
-            }
-          ]
-        })
-      }
-    })
   )
   // console.log(operations)
 
@@ -185,15 +215,6 @@ PgDb.connect().then(async pgdb => {
   clearInterval(statsInterval)
 
   if (!dry) {
-    // create tag
-    const resultTags = await fetchAuthenticated('POST', `${MAILCHIMP_URL}/3.0/lists/${MAILCHIMP_MAIN_LIST_ID}/segments`, {
-      body: JSON.stringify({
-        'name': BONUS_TAG_NAME,
-        'static_segment': []
-      })
-    })
-    console.log('mailchimp tag creation:', resultTags)
-
     const batchesUrl = `${MAILCHIMP_URL}/3.0/batches`
 
     const result = await fetchAuthenticated('POST', batchesUrl, {
